Move scroll + dfs reduce code from SearchPhaseController to actual users (#119726)
No need to have this logic live in `SearchPhaseController` when it only has a single callsite
elsewhere.

backport of #119726
