+++ server/src/main/java/org/elasticsearch/action/search/SearchScrollQueryThenFetchAsyncAction.java
@@ -66,9 +66,7 @@ final class SearchScrollQueryThenFetchAsyncAction extends SearchScrollAsyncActio
         return new SearchPhase("fetch") {
             @Override
             public void run() {
-                final SearchPhaseController.ReducedQueryPhase reducedQueryPhase = SearchPhaseController.reducedScrollQueryPhase(
-                    queryResults.asList()
-                );
+                final SearchPhaseController.ReducedQueryPhase reducedQueryPhase = reducedScrollQueryPhase(queryResults.asList());
                 ScoreDoc[] scoreDocs = reducedQueryPhase.sortedTopDocs().scoreDocs();
                 if (scoreDocs.length == 0) {
                     sendResponse(reducedQueryPhase, fetchResults);
